requirements.txt
webdriver-manager
selenium
helium
ollama
python-dotenv